requirements/install_requires.txt
Django==1.5.1
Fabric==1.5.2
FeinCMS==1.6.3
MySQL-python
pillow
South==0.7.6
django-extensions==1.0.1
django-imapauth==0.1
django-mptt==0.5.4
django-progressbarupload==0.1.6
django-rq==0.5.1
django-rq-dashboard==0.2
paramiko==1.9.0
pycrypto==2.6
python-dateutil==2.1
pytz==2012j
redis==2.7.2
rq==0.3.4
six==1.2.0
times==0.6
wsgiref==0.1.2
django_compressor==1.3
PyPDF2==1.18
reportlab==2.7
netaddr==0.7.10
scp==0.5.1
django-haystack==2.0.0
pysolr==3.0.6
django-rosetta==0.7.2
django-autocomplete-light==1.4.2
django-ressources<=0.0.4
requests==2.0.1
google-api-python-client==1.2
soundcloud==0.4.1
VideoConverter<=1.0.3
